Fixed a couple of possible stack overflows.
These would have occurred if we were trying to add a constraint to a SlotItem
that required two participating SlotItems (e.g. Pair, Incompatible), but both
participants were the same SlotItem.

This makes no sense semantically and is indicative of an error, but it still
shouldn't cause a stack overflow.