Added OTree class:
OTree can be constructed as a default with no parameters which will
initialize a new s0 evaluation. Or, a Partial assignment and unassigned
list can be specified to evaluate partial assignments. Once initialized
via the constructor, you can call genSolution() which will execute the
Or-Tree and return a new Or-Tree that should be a valid solution. If no
Valid solution could be generated, the returned Or-Tree will have its
solution value set to eSolution.NOSOLUTION which can be checked via
OTree::getSolution().

Notes: 
- Unassigned List is not initialized properly because I'm unsure where
to pull this from yet. 
- Penalties are defaulted to (0,0,0,0); Not sure what these should be
set to otherwise.
- Partial assignments aren't checked to verify that Assigned (union)
Unassigned == Courses (and) Labs.
- TimeTable is initialized with basic parameters, not sure if this needs
to be different, but Assignments::assign() requires a timetable class. 
- Some other work will need to be done as other parts of the Assignment
get finished to properly initialize and execute the Or-Tree.

Let me know if you have any questions!